backend/requirements/production.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.26.5
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
multidict==6.0.4
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.1
requests==2.31.0
rpds-py==0.13.2
sqlparse==0.4.4
twilio==8.10.3
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
yarl==1.9.3
